Create FloatValue class to hold both val and bits
This is intended to avoid some duplication between showFloat, splitSEM,
and packSEM once I write the latter. For now I save a small amount of
code in showFloat, but add more by writing the new class.

Fix a minor bug where the int10*ULP format ignored the sign of the
value.